Added laser bolt to tanks and player blaster - cross hairs always at blaster hit point
 I imported the music but I did not set it to play.  The tank can kill
the player, but bugs exist i the player killing the tank.  the player
can shoot a gree laset at the tank, but hits do not register.  If you
survive the survival time you win.  People we have a game that meets
the minimum if we can get the bugs fixed and the code commented and the
music added and the sound like you want.