Test to see if commits work for me
I added a couple of rocks/boulders to the scene and i also added a
bridge that crosses the lake. I didnt want to add too much objects in
the scene since i want to make sure that my commits work. Will be adding
more detail to our level soon.